requirements.txt
fastapi
uvicorn
docx2txt
python-docx
jinja2
textstat
